feature(app): convert web app to use ViteJS
