+++ src/components/PricesChart/index.jsx
