+++ src/pages/PrivacyTerms/index.jsx
