+++ src/services/api/index.js
@@ -1,6 +1,6 @@
 import axios from "axios";
 
-const url = process.env.REACT_APP_API_URL;
+const url = import.meta.env.VITE_API_URL;
 
 const api = axios.create({
   baseURL: url,
