+++ vite.config.js
@@ -1,0 +1,8 @@
+import { defineConfig } from "vite";
+import react from "@vitejs/plugin-react";
+
+export default ({ mode }) => {
+  return defineConfig({
+    plugins: [react()],
+  });
+};
